Finalized RepositoryTest, organized Frontend, started updating Frontend

+++ Backend/PriceTag/src/main/java/com/antunmod/pricetag/model/database/Size.java
@@ -18,7 +18,7 @@ public class Size implements Serializable {
 	@Id
 	@Column(name = "product_size_ID")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int productSizeId;
+	private Byte productSizeId;
 
 	@Column(name = "size_type")
 	private String sizeType;
